+++ packages/terra-application-utility/src/terra-dev-site/doc/example/ApplicationMenuUtilityExample.jsx
@@ -29,7 +29,7 @@ class ApplicationMenuUtilityExample extends React.Component {
 
   render() {
     const accessory = <Image alt="Fallback Avatar" src={FallbackAvatar} />;
-    const title = 'User Name';
+    const title = 'John Doe';
 
     return (
       <div>